extract quick check retrieval module
